#10: convert navigation container to not use scaffold
